+++ app/src/main/java/com/forrestguice/suntimeswidget/calendar/SuntimesCalendarDescriptor.java
@@ -23,13 +23,9 @@ import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.content.pm.PermissionInfo;
 import android.content.pm.ResolveInfo;
 import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
 import android.util.Log;
-
-import com.forrestguice.suntimeswidget.calendar.task.calendars.SolsticeCalendar;
 
 import java.util.HashMap;
 import java.util.List;
