+++ app/src/main/java/com/forrestguice/suntimeswidget/calendar/SuntimesCalendarFactory.java
@@ -21,6 +21,7 @@ package com.forrestguice.suntimeswidget.calendar;
 import android.content.Context;
 import android.util.Log;
 
+import com.forrestguice.suntimeswidget.calendar.SuntimesCalendarDescriptor;
 import com.forrestguice.suntimeswidget.calendar.intf.SuntimesCalendar;
 
 @SuppressWarnings("Convert2Diamond")
