+++ app/src/main/java/com/forrestguice/suntimeswidget/calendar/ui/AboutDialog.java
@@ -16,7 +16,7 @@
     along with SuntimesCalendars.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package com.forrestguice.suntimeswidget.calendar;
+package com.forrestguice.suntimeswidget.calendar.ui;
 
 import android.app.Activity;
 import android.app.Dialog;
