+++ app/src/main/res/xml/preference_calendars.xml
@@ -11,49 +11,49 @@
             android:summary="@string/configLabel_calendars_enabled_summary" android:defaultValue="@string/def_app_calendars_enabled" />
 
         <!--
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_astroTwilightCalendar"
             android:title="@string/calendar_astronomical_twilight_displayName"
             android:summary="@string/calendar_astronomical_twilight_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_astrotwilight" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_nauticalTwilightCalendar"
             android:title="@string/calendar_nautical_twilight_displayName"
             android:summary="@string/calendar_nautical_twilight_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_nauticaltwilight" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_civilTwilightCalendar"
             android:title="@string/calendar_civil_twilight_displayName"
             android:summary="@string/calendar_civil_twilight_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_civiltwilight" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_moonriseCalendar"
             android:title="@string/calendar_moonrise_displayName"
             android:summary="@string/calendar_moonrise_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_moonrise" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_moonPhaseCalendar"
             android:title="@string/calendar_moonPhase_displayName"
             android:summary="@string/calendar_moonPhase_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_moonphase" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_moonApsisCalendar"
             android:title="@string/calendar_moonApsis_displayName"
             android:summary="@string/calendar_moonApsis_summary"
             android:icon="?attr/icActionCalendar"
             android:defaultValue="@string/def_calendars_calendar_moonapsis" />
 
-        <com.forrestguice.suntimeswidget.calendar.SuntimesCalendarPreference
+        <com.forrestguice.suntimeswidget.calendar.ui.SuntimesCalendarPreference
             android:key="app_calendars_calendar_solsticeCalendar"
             android:title="@string/calendar_solstice_displayName"
             android:icon="?attr/icActionCalendar"
